backend/pyproject.toml
[tool.poetry]
name = "playstation-reddit-sentiment-analysis"
version = "0.1.0"
description = ""
authors = ["Elijah Khor <elijahkcm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
praw = "^7.8.1"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"
langchain = "^0.3.14"
langchain-ollama = "^0.2.2"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
